requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# Optional dependencies (uncomment as needed)

# Database adapters (if not using SQLite)
# psycopg2-binary>=2.9.0  # PostgreSQL adapter

# Authentication (when implementing Auth0)
# auth0-python>=3.24.0  # Auth0 SDK for authentication

# HTTP/API clients (for platform plugins)
# requests>=2.31.0  # For making HTTP requests to external APIs
# aiohttp>=3.9.0  # Async HTTP client (optional, for async operations)

# Development tools
# mypy>=1.7.0  # Type checking
# pytest>=7.4.0  # Testing framework
# black>=23.0.0  # Code formatting
# flake8>=6.1.0  # Linting

# Data validation (optional)
# pydantic>=2.5.0  # Data validation and settings management

